+++ altrpnjs/app/Controllers/Http/AltrpRouting.ts
@@ -48,34 +48,8 @@ export default class AltrpRouting {
     return _.get(this.__altrp_global__, path, _default)
   }
 
-  public async index(httpContext: HttpContextContract) {
-    /**
-     * Игнорим все запросы кроме get
-     */
-
-    const url = httpContext.request.url();
-    /**
-     * Игнорим логинизацию
-     */
-    for (const route of IGNORED_ROUTES) {
-      if (route === url) {
-        return
-      }
-    }
-
-    const modulesUrl = httpContext.request.protocol() + "://" + httpContext.request.host() + "/modules";
-
-    if (httpContext.request.completeUrl().split(modulesUrl).length > 1) {
-      return
-    }
+  public async getContentByUrl(url, httpContext: HttpContextContract):Promise<void>{
     const asCheck = isRobot(httpContext.request.headers())
-
-    /**
-     * Игнорим админку и ajax
-     */
-    if (url.split('/')[1] === 'admin' || url.split('/')[1] === 'ajax') {
-      return
-    }
 
     const start = performance.now();
     console.log(performance.now() - start);
@@ -204,7 +178,7 @@ export default class AltrpRouting {
 
     const altrpContext = {
       ...model_data,
-        ...pageMatch.params,
+      ...pageMatch.params,
       altrpuser,
       altrppage: {
         title,
@@ -267,6 +241,7 @@ export default class AltrpRouting {
         }
       }
 
+      httpContext.response.header('Content-Length', res.length)
       return httpContext.response.send(res)
     } catch (e) {
       console.error(`Error to View Custom Page \`${page.guid}\` : ${e.message}
@@ -298,6 +273,38 @@ export default class AltrpRouting {
          ${e.stack}
          `)
     }
+  }
+
+  public async index(httpContext: HttpContextContract) {
+
+    /**
+     * Игнорим все запросы кроме get
+     */
+
+    const url = httpContext.request.url();
+    /**
+     * Игнорим логинизацию
+     */
+    for (const route of IGNORED_ROUTES) {
+      if (route === url) {
+        return
+      }
+    }
+
+    const modulesUrl = httpContext.request.protocol() + "://" + httpContext.request.host() + "/modules";
+
+    if (httpContext.request.completeUrl().split(modulesUrl).length > 1) {
+      return
+    }
+
+    /**
+     * Игнорим админку и ajax
+     */
+    if (url.split('/')[1] === 'admin' || url.split('/')[1] === 'ajax') {
+      return
+    }
+    httpContext.response.header('Cache-Control', 'no-cache')
+    return this.getContentByUrl(url, httpContext)
   }
   async tryRenderEdgeTemplate({page,
                                 httpContext,
@@ -371,6 +378,7 @@ export default class AltrpRouting {
           }
         }
       }
+      httpContext.response.header('Content-Length', res.length)
       return httpContext.response.send(res)
     } catch (e) {
       console.error(`Error to View Custom Page \`${page.guid}\`: ${e.message}