+++ resources/modules/editor/src/js/components/altrp-posts/altrp-posts.js
@@ -225,7 +225,7 @@ class AltrpPosts extends React.Component {
     if(! this.checkStore()){
       return
     }
-    const { currentPage } = this.state;
+
     let { data: posts } = this.props;
     if (!_.isArray(posts) && _.isObject(posts)) {
       posts = [posts];
@@ -234,6 +234,7 @@ class AltrpPosts extends React.Component {
       posts = [];
     }
     let postsStart = 0;
+    const { currentPage } = this.state;
     const posts_per_page =
       Number(getResponsiveSetting(this.props.settings, "posts_per_page")) || 12;
     if (posts_per_page && Number(posts_per_page) && posts_per_page > 0) {